requirements.txt
django
djangorestframework
django-filters

requests
lxml
